requirements.txt
numpy>=1.20.1
matplotlib>=3.3.4
pyserial>=3.5
scipy>=1.8.0